pyproject.toml
[tool.poetry]
name = "recorder"
version = "0.1.0"
description = ""
authors = ["David Svenson <davidsvenson@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.6"
mistune = "^0.8.4"
click = "^7.0"
beautifulsoup4 = "^4.8"
jinja2 = "^2.10"
attrs = "^19.3"
pyyaml = "^5.1"


[tool.poetry.dev-dependencies]
black = "^19.10b0"
pylint = "^2.4"

[tool.poetry.scripts]
recorder = "recorder.cli:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
